Revert "Fix tests"
This reverts commit 69565a115175eaede5501e7c04174ee1ddde986f.
